move click functionality into a component.
