+++ Assets/Scripts/Rendering/NameRenderer.cs
@@ -17,11 +17,12 @@ public class NameRenderer : MonoBehaviour
         // Generate the initial text to be displayed above Queues.
         GameObject textObj = new GameObject("NameText", typeof(TextMesh));
         textObj.transform.parent = gameObject.transform;
+
         // This is quick and dirty hack
         // Since we have to rotate the Channel game object, so that
         // mesh collider is aligned with mesh, the Channel game object
         // is rotated and we need to position text with respect to different axis
-        if (TryGetComponent(out Channel channel))
+        if (TryGetComponent(out Channel _))
         {
             textObj.transform.localPosition = new Vector3(0, 0, 5);
         } else
